Add Group Goals file
This change adds a file to track some future goals for the usergroup.
These goals are as much thinking out load about things to do and a
personal check list as anything. That said, anyone viewing the file is
welcome to take on any of the tasks and use github issues to discuss
how best those tasks might be accomplished. If you decide to complete a
task, please do a PR against the file with the checkbox checked so that
we can use, at least in part, to keep track of who has accomplished
which tasks.